scraper/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
schedule>=1.2.0